Moved `PeopleData`, `PersonPropertyHolder`, and related structs to submodule `people::data`.
